+++ libraries/u2/src/core/widget-like.ts
@@ -1,2 +1,2 @@
-export type {ObservableLike, EventTypeLike, RectBoundsLike, InputStatusLike, WidgetStatusLike, FuncLike,
+export type {ObservableLike, IEventType, IRectBounds, IInputStatus, IWidgetStatus, FuncLike,
   WidgetLike, BindPropLike, BindSourceLike, ComponentMetaLike} from 'datagrok-api/u2core';